test: use pyreltrans to run the tests
It is significantly faster in the CI to use the pyreltrans wrapper
instead of the HEASoft Docker container, hence the ci and tests are
being migrated.

This commit currently does prevent running the tests without pyreltrans
installed, but we can modify the pyreltrans wrapper to accept an
arbitrary path the reltrans library such that it defaults to whatever is
in the build directory and falls back to the Python library path.
