+++ .github/workflows/test.yml
@@ -34,16 +34,20 @@ concurrency:
 jobs:
   test:
     runs-on: ubuntu-latest
-    container: fjebaker/heasoft:ci-6.33.1-amd64
     steps:
       - name: Checkout
         uses: actions/checkout@v5
+      - name: Set up Python
+        uses: actions/setup-python@v6
         with:
-          fetch-depth: 0
-      - name: Build debug reltrans
+          python-version: "3.10"
+      - name: Install dependencies
         run: |
-          make DEBUG=1
-          make DEBUG=1 exe
+          python3 -m pip install --upgrade xspectrampoline numpy pytest
+      - name: Compile pyreltrans and install
+        run: |
+          python3 ./dist-package.py
+          python3 -m pip install .
       - name: Install GitHub CLI and fetch data files
         run: |
           apt update && apt install -y gh
@@ -58,11 +62,7 @@ jobs:
           # Stitch the large files back together
           cd reltrans-data
           cat $(ls xillverCp_v3.4_normalised.fits-* | sort -V) > xillverCp_v3.4_normalised.fits
-      - name: Setup pytest environment
-        run: |
-          python -m pip install --upgrade pytest numpy
-      - name: Run tests
+      - name: Smoketest pyreltrans
         run: |
           export RELTRANS_TABLES="$(realpath reltrans-data)"
-          export RELTRANS_PATH="$(realpath build/lib/libreltrans.so)"
           pytest -vvv test
